+++ DreamCalendar/DreamCalendar/Scene/ScheduleAddition/SchduleAdditionView.swift
@@ -1,0 +1,15 @@
+//
+//  SchduleAdditionView.swift
+//  DreamCalendar
+//
+//  Created by 이지수 on 2022/11/19.
+//
+
+import SwiftUI
+struct ScheduleAdditionView: View {
+    var body: some View {
+        VStack {
+            Text("일정 등록")
+        }
+    }
+}
